Fix init seats for auditoriums. Refactored strategies, services, dao layers

+++ src/main/java/ua/epam/spring/hometask/service/BookingService.java
@@ -6,6 +6,7 @@ import ua.epam.spring.hometask.domain.User;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Set;
 
@@ -25,8 +26,8 @@ public interface BookingService {
      * @param seats    Set of seat numbers that user wants to buy
      * @return total price
      */
-    double getTicketsPrice(@Nonnull Event event, @Nonnull LocalDateTime dateTime, @Nullable User user,
-                           @Nonnull Set<Long> seats);
+    BigDecimal getTicketsPrice(@Nonnull Event event, @Nonnull LocalDateTime dateTime, @Nullable User user,
+                               @Nonnull Set<Long> seats);
 
     /**
      * Books tickets in internal system. If user is not
